% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lh-m-chenWatanabe.R
\docType{methods}
\name{chenWFn}
\alias{chenWFn}
\title{chenW}
\usage{
chenWFn(age, par)
}
\arguments{
\item{age}{age at which M is to be predicted}

\item{par}{FLPar or numeric vector  with von Bertlanffy parameters \code{t0, linf, k}

#'}
}
\description{
Chen and Watanabe natural mortality equation
provides an estimate of M per year at age as a function of von Bertalanffy growth equation
}
\examples{
\dontrun{
par=FLPar(linf=200,k=.5,t0=-.3)
chenW(1:10,par)
}
}
\seealso{
\code{\link{chenW},\link{gislason}}
}

