% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{lorenzen}
\alias{lorenzen}
\title{chenWatanabe}
\usage{
lorenzen(wt)
}
\arguments{
\item{par}{FLPar with von Bertlanffy parameters \code{t0, linf, k}}

\item{age}{for which growth to be predicted

#'}
}
\description{
Chen and Watanabe natural mortality equation
}
\examples{
\dontrun{
par=FLPar(linf=100,t0=-0.1,k=.4)
age=FLQuant(1:20,dimnames=list(age=1:20))
m  =chenWatanabe(par,age)
}
}
\seealso{
\code{\link{invVonB}} \code{\link{gompertz}}
}

