% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{gislasim}
\alias{gislasim}
\title{gislasim}
\usage{
gislasim(par, t0 = -0.1, a = 1e-05, b = 3, ato95 = 1, sl = 2,
  sr = 5000, s = 0.9, v = 1000)
}
\arguments{
\item{\code{par}}{an \code{FLPar} object with life history parameters}

\item{\code{growth}}{function for growth}

\item{\code{fnM}}{function for natutal mortality}

\item{\code{fnMat}}{function for proportion mature-at-age}

\item{\code{fnSel}}{function for selectivity-at-age}

\item{\code{sr}}{character for stock recruitment relationship}

\item{\code{range}}{age range}

\item{\code{spwn}}{proportion of year when spawning occurrs, i.e. level of natural mortality prior to spawning}

\item{\code{fish}}{proportion of year when fishing happens}

\item{\code{units}}{}

\item{...}{}
}
\value{
\code{FLBRP}
}
\description{
Takes an \code{FLPar} object with life history and selectivity parameters
and generates an corresponding \code{FLBRP} object. Can uses a range of functional forms
}
\examples{
\dontrun{
data(bd)
bd=fit(bd,swonIndex)}
}
\seealso{
\code{\link{vonB}} \code{\link{lorenzen}} \code{\link{logistic}} \code{\link{doubleNormal}}
}

