% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lh-growth-gompertz.R
\docType{methods}
\name{gompertz}
\alias{gompertz}
\title{gompertz}
\usage{
gompertz(par, age)
}
\arguments{
\item{par}{FLPar with parameters for \code{linf, a, b}}

\item{age}{#'}
}
\description{
Gompertz growth equation
}
\examples{
\dontrun{
par=FLPar(linf=100,t0=0,k=.4)
age=FLQuant(1:10,dimnames=list(age=1:10))
len=gompertz(par,age)
}
}
\seealso{
\code{\link{vonB}}
}

