% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lh-OM.R, R/lh-func.R
\docType{methods}
\name{omFn}
\alias{lh}
\alias{omFn}
\title{om}
\usage{
omFn(x, b0 = 0.8, b = 0.1, year = seq(40, 90), maxF = 5)

lh(par, growth = lh:::vonB, fnM = function(par, len) exp(0.01 * par["m1"]
  \%-\% (log(len) \%*\% par["m2"])), fnMat = logistic, fnSel = dnormal,
  sr = "bevholt", range = c(min = 1, max = 40, minfbar = 1, maxfbar = 40,
  plusgroup = 40), spwn = 0, fish = 0.5, units = if ("units" \%in\%
  names(attributes(par))) attributes(par)$units else NULL, ...)
}
\arguments{
\item{\code{par}}{\code{FLPar} object with parameters for life history equations and selection pattern.
Need L_infty to estimate other parameters, if any other parameters supplied in \code{code} then
these are not provided by the algorithm}

\item{\code{t0}}{of von Bertalanffy. This is a default that isnt normally derived
from life history theory, as are the following args.}

\item{\code{a}}{coefficient of length weight relationship}

\item{\code{b}}{exponent of length weight relationship}

\item{\code{ato95}}{age at which 95\% of fish are mature, offset to age at which 50\% are mature}

\item{\code{sl}}{selectivity-at-age parameter, standard deviation of lefthand limb of double normal}

\item{\code{sr}}{stock recruitment relationship}

\item{\code{s}}{steepness of stock recruitment relationship}

\item{\code{v}}{virgin biomass}

\item{\code{par}}{\code{FLPar} object with parameters for life history equations and selection pattern.
Need L_infty to estimate other parameters, if any other parameters supplied in \code{code} then
these are not provided by the algorithm}

\item{\code{t0}}{of von Bertalanffy. This is a default that isnt normally derived
from life history theory, as are the following args.}

\item{\code{a}}{coefficient of length weight relationship}

\item{\code{b}}{exponent of length weight relationship}

\item{\code{ato95}}{age at which 95\% of fish are mature, offset to age at which 50\% are mature}

\item{\code{sl}}{selectivity-at-age parameter, standard deviation of lefthand limb of double normal}

\item{\code{sr}}{stock recruitment relationship}

\item{\code{s}}{steepness of stock recruitment relationship}

\item{\code{v}}{virgin biomass}
}
\value{
An \code{FLPar} object with parameters

An \code{FLPar} object with parameters
}
\description{
Uses life history theory to derive parameters for biological relationships,
i.e. growth, maturity, natural mortality from.

Uses life history theory to derive parameters for biological relationships,
i.e. growth, maturity, natural mortality from.
}
\examples{
\dontrun{?om}
\dontrun{
lh(FLPar(linf=200))
}
}

