% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lh-ogive-logistic.R
\docType{methods}
\name{logistic}
\alias{logistic}
\title{logistic}
\usage{
logistic(params, data, ...)
}
\arguments{
\item{par}{}

\item{age}{#'}
}
\description{
logistic function
}
\examples{
\dontrun{
par=FLPar(linf=100,t0=0,k=.4)
age=FLQuant(1:10,dimnames=list(age=1:10))
mat=logistic(par,age)
}
}
\seealso{
\code{\code{\link{gompertz}}}
}

