% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{vonB}
\alias{chenWatanabe}
\alias{vonB}
\title{vonB}
\usage{
vonB(par, age)

chenWatanabe(par, age)
}
\arguments{
\item{par}{}

\item{age}{#'}

\item{par}{FLPar with von Bertlanffy parameters \code{t0, linf, k}}

\item{age}{for which growth to be predicted

#'}
}
\description{
Von Bertalanffy growth equation

Chen and Watanabe natural mortality equation
}
\examples{
\dontrun{
par=FLPar(linf=100,t0=0,k=.4)
age=FLQuant(1:10,dimnames=list(age=1:10))
len=vonB(par,age)
age=invVonB(par,len)
}
\dontrun{
par=FLPar(linf=100,t0=-0.1,k=.4)
age=FLQuant(1:20,dimnames=list(age=1:20))
m  =chenWatanabe(par,age)
}
}
\seealso{
\code{\link{invVonB}} \code{\link{gompertz}}

\code{\link{invVonB}} \code{\link{gompertz}}
}

