% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lh-m-jensen.R
\docType{methods}
\name{jensen}
\alias{jensen}
\title{jensen}
\usage{
jensen(params, data = NULL)
}
\arguments{
\item{par}{FLPar with von Bertlanffy parameter \code{k}

#'}
}
\description{
Jensen natural mortality estimate based on k of the von Bertalanffy growth equantion
}
\examples{
\dontrun{
jensen(FLPar(k=.5)
}
}
\seealso{
\code{\link{lorenzen}}
}

